Refactor gravitational force into new module
